TraceReader: Provide TraceReader concept and a basic reader to read lines
The TraceReader is a packetizer that reads arbitrary streams and converts to lines. These lines decode particular packets, be they strings, or a break down of a complex multiplexed protocol.

See `docs/Architecture.md` for design of the implementation and how the classes are organized.

A basic reader is provided, that can read text files, and provides individual lines that are terminated with a new line character (the `TextTraceReaderFactory`). A new line is defined as any line that is terminated with either 0x0a (Unix / Darwin), 0x0d 0x0a (Windows), 0x0d (Classic MacOS).

Issue: DOTNET-477