DLT: Implement decoder for DLT TCP streams
This commit implements the minimum part for decoding a DLT trace message. It's not yet useful enough to do anything, except to parse the ECU ID, Application ID, Context ID and the message type information.

A white paper is written also `DLT.Format.Problems.md`, which describes the DLT format, and the problems that decoder implementations face with DLT streams. It also offers a potential solution that may be compatible to existing implementations.

Issue: DOTNET-484