DLT: Implement Decode.Flush
Flushing a stream (parsing the data that is already in the cache) is in almost all cases the same as decoding a normal stream, except we know that there will not be more data, so nothing should be cached at the end of the operation.

The Flush() method the calls decode on whatever data is in the cache. In Debug mode, we try to avoid bugs by ensuring any access to the cache within the Decode method results in an ArgumentNullException. This isn't done in Release mode as a performance optimization.

New test cases added show what happens when there is data present in the buffer when flushing. It is generally converted to a new line indicating skipped data.

Issue: DOTNET-497