DLT: Use real buffer if possible
Try to use data from the input buffer instead of the line cache where possible. It could occur that data is cached through scanning (a previous message is incomplete, so it is cached), and then it is considered invalid, there is a lot of data in the cache. If it happens that this data in the cache is also in the input buffer, we should use the input buffer instead.

By using the input buffer instead, we allow for fewer operations when scanning for the marker, and potentially avoid a lot of future copy operations.

There was no notable change in the performance tests.

Issue: DOTNET-499