DLT: Set up tracing infrastructure
The next steps are to implement the decoder. There will be places that decoding will detect invalid data. No exceptions should be raised so that the decoder is as performant as possible. But it should be able to log, and it will use the RJCP.DLL.Trace library to do this.

Issue: DOTNET-488