DLT: Extend IDltLineBuilder to build verbose arguments
When a decoder parses the payload, it will use the IDltLineBuilder to add each parsed argument.

There is a slight performance impact when decoding, that has to do with the construction of the `DltTraceLine` where the argument list is copied by the constructor.

Code coverage (VS2019 IDE) at 92.9%

Issue: DOTNET-501