DLT: Provide decoders for Verbose Types (Int, Float, Bool, String, Raw)
Various decoders are provided that decode verbose payloads. The payload starts at the first type where the type information is stored.

The decoders are:
- Signed Integer (128-bit returns unknown argument)
- Unsigned Integer (128-bit returns unknown argument)
  - with Hex and Binary coding (not in AutoSAR)
- Boolean
- Float (16-bit and 128-bit return unknown argument)
- Strings (ASCII = ISO-8859-15 and UTF8)
- Raw Arguments

The Unknown Argument was introduced, where the size of the payload is known, but there is no reasonable representation in .NET. So the base representation is an array which is copied. This is used for 128-bit integers for example. It is different from not supporting the argument, which would render the entire packet as invalid.

The VerboseArgDecoder looks at the type information to determine which ArgDecoder should be used. Test cases shouldn't be able to discern between using the ArgDecoder and the VerboseArgDecoder if it can be properly decoded.

Issue: DOTNET-504