DLT: Decode verbose messages in the main decoder
After decoding the message, if the payload length doesn't match the length expected, heuristics treats this packet as invalid and skipped data will result.

Simplify test cases by defining "common" messages that can be tested against.

This implementation relies on .NET doing boundary checks with the Span<T> type. If there is corruption in a packet, boundaries may be exceeded which will be caught by the runtime.

Issue: DOTNET-505