DLT: Skipped lines shouldn't support ECUID, DeviceTimeStamp
When we detect a skipped line, it's because data is corrupt. Don't make assumptions about data received over the wire, which means that the ECUID and the Device Time Stamp should be considered not available and be set to appropriate defaults.

Only the Time Stamp as received by the logger might be considered valid.

Issue: DOTNET-506