DLT: Provide a dedicated skipped line
By providing a line that is explicit for skipped bytes, it can be easier to later gather statistics and heuristics on the amount of corrupt data in a DLT file without having to parse the contents of the line as strings.

Test cases were updated to check the correct number of skipped bytes.

Issue: DOTNET-506