DLT: Fix decoding exception when new packet is already in the cache
When data is already cached through a previous decode operation, and a new packet must be decoded where that new packet is already completely within the cache, the decoder would crash with an ArgumentOutOfBounds exception. Protected against this case.

This was found when adding a test case decoding random data. It could occur with correct data if a packet is corrupted with a very large length, and a second packet can be decoded that is already within the cache.

Issue: DOTNET-506