DLT: Optimize construction of DltTraceLine arguments
Allow the DltTraceLine to have an argument list that is fixed when created from an array (the DltLineBuilder). Create a new `IList<IDltArg>` implementation (called `ReadOnlyArrayList<T>`) that wraps around the array, to copy only the array reference, instead of trying to enumerate and copy.

This provides a very minor performance improvement for the construction of each line, and makes the argument list effectively immutable.

Issue: DOTNET-502