DLT: Implement decoding GetSoftwareVersion request and response in main decoder
Provide the `IControlDltDecoder` which is given to the main decoder, and this looks up the service identifier to find the appropriate control payload decoder. There is only one control payload decoder implemented in this version.

Testing is done for the individual decoder, from the `IControlDltDecoder` and from the main decoder.

The micro-benchmarks show a small hit to decoding as the packet is now checked if it is a control message or not, with about 3ns per message (30ns for the 10 messages being measured).

Issue: DOTNET-515