DLT: Add control message SwInjection
All service ID's of 0xFFF..0xFFFFFFFF are valid requests to software with a payload.

It is possible to derive from the object `SwInjectionRequest` and the decoder `SwInjectionRequestDecoder` to provide more specific implementations for a software injection callback. To provide this callback, one must also derive from ControlDltDecoder, register the callback in the constructor, and then inject into the trace decoder.

As the range of control messages which are now valid is very high, the test case `DltTraceDecoderCommonTest.RandomData` is likely to find a valid packet within random data, thus no longer skipping as many bytes as it did prior. The test case no longer checks the number of skipped bytes, as we can't calculate any more all data that was decoded. The test case just makes sure that fuzzing doesn't result in crashes.

Issue: DOTNET-520