DLT: Provide decoding of Big-Endian Verbose Arguments
Many arguments are written already to handle the encoding, but the main encoder was always treating the DLT Type as little endian, where it should be dependent instead on endianness flag of the packet. This change is quite minor, with the majority being the test code required to test.

Take a similar approach to how control messages are tested, and allow writing each verbose argument to be tested to disk, that it can be independently loaded with other tools and checked for consistency (including display).

Now that the VerboseDltDecoder takes the typeInfo, the exception handling needs to be moved there.

The micro-benchmarks show that decoding the verbose arguments is slightly faster, because it is now given the Type Info field and just needs to apply masks on this. The Type Info field is calculated from another class as part of the main decoder. The overall processing for integers and strings is slightly slower as the full type info is now calculated to handle the endianness properly.

Issue: DOTNET-539