DLT: Decode non-verbose data as byte strings
With the introduction of a simple (and very accepting) non-verbose decoder, the DltType.UNKNOWN is now a valid DLT type, when there is a non-verbose payload that is present. This information would then normally come from the FIBEX file.

Most of the changes are refactoring of test code to generate the appropriate datastreams to test the decoding of non-verbose. Created the new DecoderTestBase and have ControlDecoderTestBase, VerboseDecoderTestBase and the new NonVerboseByteDecoderTestBase derive from this as they all write samples that can be loaded with the DLT-Viewer for comparison.

The random tests are now significantly faster, that the probability of a non-verbose message is now higher. Thus the test case was extended with more data and repetition to try and fuzz for errors.

Issue: DOTNET-511