DLT: Fix a critical bug that the DltLineBuilder wasn't properly finding control messages
A control message was being created whenever bits 1 or 2 of the DltType might have been set, which is for CONTROL messages, but also for other message types. We should be masking for the control message first and then checking.

This means that network messages (0xX2 or 0xX4, e.g. NW_TRACE_ETHERNET = 0x54 & 0x06 != 0) could have been interpreted as control messages, but not normal trace messages

Issue: DOTNET-511