DLT: Provide position for each packet generated
Maintain a position map, that is a list of blocks with a position and length. This position map is updated every time data is added or removed from the stream of data to decode. It is possible that the data to decode has gaps (e.g. the input stream is packet based), and as data is consumed, and DLT packets jump over these gaps, the position of the start of the packet tracks this.

This has about a 4% performance impact to track the position. This has a large benefit in tracking down the location of a packet when errors occur.

The version compiled in DEBUG mode will check additional invariants to ensure no logical errors while decoding. These exceptions are disabled in Release mode, which may lead to undefined behaviour in the calculations of the position.

Issue: DOTNET-486