DltDump: Implement the version command
On Windows, `dltdump /version` shows the version and the version of the assemblies.

On Linux, `dltdump --version` shows the version and the version of the assemblies.

Test cases show that the command works. Other test cases prior have shown already that the options are parsed and the application is started.

Issue: DOTNET-555