DLT: Apply valid packet heuristics on the message type
The message type has only a range of specified bits. For example, the MTIN of zero is not specified, and only some enumeration values are specified for MTSP and MTIN.

Placing stricter heuristics on the message type can help detect corrupt packets due to byte loss over serial.

Issue: DOTNET-574