DLT: LineCache now has a "flush" mode
When flushing, put the `LineCache` in a flush mode, that raises exceptions when the cache is used, as well as clearing the cache and returning a copy of the data in the cache.

This better reflects the decoder state, and the `flush` variable is then used for where specifically different logic is needed in flushing, not necessarily which buffer to use, also keeping closer to the original design.

Issue: DOTNET-576