DLT: Print the number of arguments in the string output
The dlt-viewer shows the NOAR field when it exports the data to text. This implementation doesn't show the NOAR field, but instead the number of decoded arguments. The main difference will be for non-verbose messages, the NOAR should be zero, but this implementation will print the number of arguments that the line has.

Issue: DOTNET-577