DltDump: Implement a TcpClientStream
A `TcpClientStream` can be used to connect to a remote server. It supports application level timeouts if data is not received within a period of time. This is useful for embedded devices when they simply power off, to avoid excessive time outs due to the TCP state machine.

Issue: DOTNET-580