DltDump: Apply option /retries for connection retries
If the retries is more than zero, the application will loop forever, until all the retries are expired. This allows an automated connection until the number of retries are exceeded. If the number of retries is set to be a negative number, it will retry infinitely, so the application will never end on its own.

Issue: DOTNET-582