TraceReader: Add Constraints
Constraints is a way to efficiently check the fields of a line for a match of data. It uses a fluent style of joining methods, that results in a binary expression tree being built that then leads to IL generation to do the checks as fast as possible. This code is useful for checking regularly data from the `ITraceLine` in a tight loop.

Issue: DOTNET-565