DltDump: Update design for output to files
Writing the output should consider various use cases on how to concatenate files, but also in complex scenarios which depend on the time stamp. The context and filter can be combined into a logical write operation, which can be chained to something that writes a specific format, be it text or DLT. It could theoretically be used to convert from DLT to any other format so long as it is known how to write that based on the actual line and/or byte packet information.

Issue: DOTNET-592