DltDump: Prepare factory for creating output files
This first implementation sets a factory for creating output streams. The first output stream is only to the console, writing to files is currently not supported.

If when creating the output fails, a new error code ExitCode.OutputError (5) is returned.

The `OutputFormat` is always set to be `Automatic` when writing. This could change in the future to be provided on the command line for more complex scenarios.

Issue: DOTNET-595