DltDump: Provide a class to write text output
The `OutputWriter` is the class that can write to a file and maintains its length.

The `OutputBase` will maintain the operations required when writing to a file. At the moment it writes to a file name, and can be told to force overwrite. For the lifetime of the object, any writes output to the file. It has three write operations, that can write a string, or a buffer, or a buffer with a header, which are the three use cases when writing DLT lines. In the near future, it should also handle the output file name as a template, substituting the input file name, the date/time and splitting with a counter.

The `TextOutput` relies on the `OutputBase` to write strings as UTF8 to the underlying file.

Issue: DOTNET-597