DltDump: Allow writing to a text file
The user can now provide the option `/output:textfile.txt` and the output will be written to that text file. The option `/showposition` can be used to prepend the position within the stream for the output.

The output is consistent and printed in UTF-8. This is more reliable than trying to redirect the output on the console to a file, which under Windows depends on the command shell being used (Power Shell outputs in UCS2, Windows CMD outputs in the current code page, MSys outputs probably in UTF8).

Issue: DOTNET-598