DltCount: Quick Test Program for counting packet sizes
A small test program that prints out the count of packet sizes, which can be useful for printing statistics.

The output is in the format:

   <timestamp> <ecuid> <appid> <ctxid> : <length>

So you can count the number of lines with

   $ dltcount.exe file.dlt > out.txt
   $ wc -l out.txt

You can ensure that all bytes are accounted for with

   $ dltcount.exe file.dlt > out.txt
   $ awk -F' ' '{sum+=$7;}END{print sum;}' out.txt

The output should match the size of the file.

Issue: DOTNET-614