DltDump: Design of output files improved on when using templates
Went into more detail with the design and provided pseudocode which has resulted in a simplified implementation.

If the output template depends on the input file name, via %FILE%, we don't consider concatenation, as the output files are dependent on the input file name. Ignoring splits, this is the same as saying for each input, generate an output.

If the output template doesn't depend on the input file name, then we consider all the files should be written to one output. Ignoring splits, this is the same as just joining everything together.

When we split, the result is the same as if we split the current output. So the design talks about "Segments", which is conceptually the same just as a single output, that a Segments list is just a list of files, when concatenated together, forms the right output. This allows the algorithm to treat splitting files different to joining or concatenating, making things easier and probably more logical for the user.

Issue: DOTNET-599