DltDump: Implement Split logic
Take the option to split and apply it to the OutputBase. The option /split is used to define the size the output file must exceed when it is split. The output option /output must include a template that describes the file name, either with %CTR% (which is a counter that increments), or with %CDATE%, %CTIME% or %CDATETIME%.

Issue: DOTNET-602