DltDump: Don't overwrite input files or files converted by this tool
When converting files, we check that they all exist, and get file system information that ensures we will not overwrite the inputs. When a file is written, this is also added so if there is a name clash when overwriting we don't overwrite.

This avoids possible confusion and a collection of split files being partially overwritten, as well as overwriting an input, which may be later opened and thus have different data to the time it was opened.

Issue: DOTNET-615