DLT: Allow an object to override and get information about each packet
This will allow a `DltOutput` object to receive the packet data, that it can write it in the context of the decoder thread.

Performance benchmarks show no discernible difference when run.

The first decoder to use this information is `DltFileTraceFilterDecoder`, that overrides the `CheckLine` so that on decoding it writes to the `IOutputStream`.

Issue: DOTNET-611