DltDump: Decode PCAP legacy packets
A PCAP decoder, that extracts DLT packets from UDP packets given a PCAP file is implemented. This is implemented in the DltDump application, as it's specific to that application (it's not generic enough to put in a reusable library and will never be given the complexities of IP protocols).

The DltTraceDecoderBase has a new method that can "flush" output, which makes it more efficient to decode input packet streams, like that from a PCAP.

Issue: DOTNET-622