DLT: PosMap shouldn't require consecutive blocks
When we get to recombine fragmented IPv4 packets, the buffers will be at various positions in the input stream, and those buffers won't even be consecutive. It's perfectly fine that the last block of a fragmented packet is first, followed by the earlier packets last.

Thus remove the (debug) restriction that packets must be consecutive, and add a test case ensuring this functionality is met.

Issue: DOTNET-623