DltDump: Fix test case for VersionApp
With the commit b8d7288 ("DltDump: Improve loiteral strings and replace with resources") the VersionApp was modified to wrap the output string. In case it was more than 80 characters (which is the case when an alpha build and also dirty), it would be split into two and would fail.

Issue: DOTNET-627