DltDump: Fix out of bounds access on decoding PCAP files
In case that an IPv4 packet has a larger than normal Internet Header, and the input file is corrupted so that the packet is 50 bytes or larger, but cut so the UDP port is not available, the program would crash with an out of bounds exception.

e.g. if the first byte after the Ethernet header and prototype 0x0800 is not 0x45 (20 bytes header) as is normal, but is 0x49 (36 byte header) and the buffer is only 50 bytes, an exception would occur.

Issue: DOTNET-640