DltDump: Use only the capture size to get the packet
It has been observed that for some captures the Capture Length and the Original Length do not match. But when reviewing the packet, it still contains a complete IPv4 packet (and some times with trailing data). Thus, we should rely only on the PacketDecoder that it has enough data as per the IPv4 packet, not the captured packet size.

Issue: DOTNET-639