DltDump PCAP: Handle each src:port, dest:port as a virtual stream
Treat virtual connection, which is the pair (source:port, destination:port), as an independent stream. If a device that is logging wants to log a DLT packet that is split across multiple UDP packets, it can do so and the virtual connections won't interfere with each other, even if they interleave.

This change makes decoding DLT via UDP more sensitive to dropped packets for a virtual connection, so it is now assumed that a DLT packet may not fit entirely within a single UDP packet.

A DLT packet split across multiple UDP packets means specifically that the DLT payload is split into multiple UDP packets that themselves may not necessarily (and probably also not) fragmented.

Issue: DOTNET-625