DltDump: Handle the case of a corrupted file
In case of a corrupted file, the `BlockReader` would return `null`. This condition should be checked, and a corrupted file should be reported.

When corruption is detected, it now flags the file, continues decoding. We don't raise the exception immediately, as there may be some packets in the buffer that were decoded. This fix will then return more data. The exception is now postponed to the next `Decode()` or `Flush()` method call.

The TraceReader now treats a `null` for an enumeration to indicate that decoding can no longer continue, instead of treating it the same as an empty collection. This could mean a safe way to indicate no more decoding is required, even though the stream hasn't finished yet.

Issue: DOTNET-642