DLT: NonVerbose should be its own object
The implementation where the first argument contains the message identifier will not work when arguments are decoded using external input, because the non-verbose argument will no longer be present. Create a new type `DltNonVerboseTraceLine` that contains the argument explicitly, and allow the `DltLineBuilder` to create this object. The message identifier is removed from the Non-Verbose argument.

When creating the `DltNonVerboseTraceLine`, noticed that `DltTraceLine` generation could be optimized, and further needed to use `Text` for `DltNonVerboseTraceLine` to work.

The `DltType == DltType.UNKNOWN` is valid for the non-verbose messages, as there might not be an external file to decode.

As a minor detail, `RawDltArg` was updated to be closer to `NonVerboseDltArg`, although they're not the same when generating strings.

Issue: DOTNET-629