DltDump: Enable flushing periodically the file written to disk
Flushes to disk are for live connections, e.g. with the serial port, or via the network. When parsing text files, flushing should not occur.

Tested manually on the command line that flushing works with the command below. The code was slightly modified so that an output was printed to the console when flushed. The split is required to ensure that flushing is reestablished when a file is closed and reopened.

```sh
dltdump.exe /o:out_%CTR%.dlt /split:500M /retries:5 tcp://127.0.0.1
```

Issue: DOTNET-609