DLT: Implement Receiver PoC
Show how to receive UDP packets from a multicast group. Experimenting with this code helps to see the core of how to receive UDP packets on .NET (and can be used to verify functionality on different Operating Systems).

The test tools allow address reuse, so that the beacon and the receiver can be used on the same machine. Multiple receivers can run on the same machine.

Issue: DOTNET-647