DltDump: Recover from duplicate fragmented frames in PCAP
Some PCAP recordings can have duplicate packets, depending on the equipment for tracing. While this patch set doesn't attempt to remove duplicate packets (use WireShark's editcap first with the `-D` and `-d` options), it at least doesn't treat them as errors when reassembling packets, discarding duplicates.

Duplicates are determined as two packets having the same fragmentation offset and fragmentation identifier, and the same checksum of the Internet header block as received. It doesn't try to calculate hashes itself for the content, relying on the accuracy of the packet, for faster processing.

This of course doesn't remove duplicates for packets that aren't fragmented, only those that are. You may see some warnings for packet errors still in logs, as the last packet which is a duplicate after a frame is fragmented might be considered a new fragmented packet.

Internally, the idea of a duplicated last frame is removed, and replaced with any frame that might be a duplicate. That frame is then ignored, rather than discarding all previous packets and attempting new.

A warning about using array buffers directly was removed by replacing it with a function call (`IpFragments.GetArray()`).

Issue: DOTNET-653