DltDump: Implement UdpPacketReceiver to get UDP packets
The `IPacket` interface now defines how to read a packet, e.g. which would be a UDP packet. The `UdpPacketReceiver` implements this interface, and will bind to a single UDP address (either a multicast or a local interface), and will keep account of the end points.

The implementation of `ReceiveAsync()` doesn't wait for tokens as there is no easy way to abort a UDP packet in .NET Core 3.1.

Issue: DOTNET-652