DltDump: Properly handle file names for non-file connections
There was an error in the way that the %FILE% was handled for non-file type inputs, it would try and convert it to a file name. Now, non-file connections are explicitly treated as having now file name and is reflected in the output.

The `IInputStream` now has a property called `InputFileName` which is null, except for files. When calling `IOutputStream.SetInput`, if it is `null`, it is treated as having no file name. The `DltSerialStream`, `DltTcpStream` and `DltUdpPacketReceiver` all return `null`.

Issue: DOTNET-652