DltDump: Separate reader factories into decoder factories
The basic `*TraceReaderFactory` no longer instantiates the decoder direct. It instead creates a similar `*TraceDecoderFactory` and gives it to the base class `TraceReaderFactory` which instantiates, instead of being in the derived parent. This allows reuse of the decoder factories where the reusable logic for supporting multiple decoders reside without business logic (i.e. knowing where to receive the data, separating concerns). The documentation in the core assembly was updated to reflect this.

The `Dlt(|Serial|File)TraceReaderFactory` for creating the decoders in the TraceReader.Dlt assembly is now moved to new classes `Dlt(|Serial|File)TraceDecoderFactory`.

Renamed `IDltTraceReaderFactory` to `IDltDumpTraceReaderFactory` to make it more explicit that this isn't a specific `TraceReaderFactory` for a DLT protocol, but the factory for all protocols. TODO Documentation is added to explain the next steps for this feature in the code for `DltDumpTraceReaderFactory`.

The `DltDumpTraceReaderFactory` uses a new `DltDumpTraceDecoderFactory` decoder factory and gives it to the `TraceReaderFactory`. This allows us to provide all of DltDump's custom decoders in one place, allowing removal of the reader factories (which were only there as wrappers around the decoders).

Issue: DOTNET-652