DltDump: Set the output properties outside the "input" loop
The `IOutputStream` must know the format of the input before a line is written (e.g. the `DltOutput` uses the input format to know if a storage header should be added or not). The `IOutputStream` requires that every `Write()` call is from a file of the same format, i.e. it doesn't allow writing from multiple different binary formats at the same time (to do that would require giving the format on each call to `IOutputStream.Write`).

So we can move the logic for setting the output format before opening the input. This will allow a later commit for opening multiple inputs of the same format simultaneously, which will be merged into the same output.

Issue: DOTNET-652