UdpPacketReceiver: Copy received data into the buffer
.NET Standard 5.0 and earlier `UDP.ReceiveFrom()` doesn't have a method to receive data giving a `Memory` structure. The original implementation was incorrect that it copied the `Memory buffer` to the array, thinking that the array and `Memory` pointed to the same region in memory (they didn't, the array is a copy). This resulted in reading data from the UDP ports into the array, but not returning it via the `Memory` structure. Code then got the right length, but the buffer was full of zeroes.

Use MSDocs in .NET 5.0 and earlier to convert the `Memory` to an `ArraySegment` and then an array. Updated test cases to check the contents of the buffer returned is correct (and that the original implementation was wrong).

The code is complicated, and can be cleaned up when we upgrade to a newer version of .NET that can support this enhancement. An example of how this works is implemented, for later reference (but should use the Async methods instead).

Issue: DOTNET-652