DltDump: Implement TracePacketReader
The TracePacketReader and its factory TracePacketReaderFactory know how to read from an IPacket, in addition as a file and a stream for handling UDP inputs with multiple endpoints.

Add events to IPacket when a new channel is seen. This allows efficient instantiation of the decoder without having to always check the channel on each line in the `TracePacketReader` (which we do anyway when we receive the binary packet in `UdpPacketReceiver`).

Issue: DOTNET-652