dltudpbeacon: Add clang-tidy linter to the build
In order to see how to improve the C++, add the clang-tidy linter.

Works on Linux after it is installed. if clang-tidy is not installed, then builds work as they did before.

For the QNX 7.1.0 cross toolchain, we disable the usage of clang-tidy, as this causes errors when building.

Issue: DOTNET-669