dltudpbeacon: use std::vector
Using a std::vector removes all the pointer arithmetic.

The std::vector is much more flexible over a std::array and the code is much more compact. The original reason for using an array was allocation on the stack to try an avoid an allocation (but 64kB on the stack is already quite large). As such, we allocate the std::vector in the constructor and it remains for the lifetime of the `dlt` class, achieving something similar. It incidentally reduces an allocation by not having to create a new vector to just send the data via UDP.

Issue: DOTNET-669