DltDump: Support Inner/Outer VLANs
Refactor the PacketDecoder to scan for protocol identifiers and process the packet based on that. This will be useful when processing TECMP files, that have a protocol of 0x99FE (IPv4 has 0x800). Later, we can reuse the code for decoding the wrapped IPv4 packet in the TECMP file.

As an addition, it's easy to now process an inner/outer VLAN (we only processed a single VLAN). We don't do anything with the VLAN, other than just strip it out.

Issue: DOTNET-695