DltDump: AutoFlush might not flush automatically at the end
Observed through a compiler warning that the method `OutputWriter.AutoFlush` has a variable that isn't used. Instead of accessing `m_FileStream` in the loop which may be set to `null` on close, grab a reference and use that instead. It will still raise an exception when the stream is closed, but not a null pointer.