DLT: Non-verbose Structures
Provide the internal datastructures that map from a message identifier, application and context identifier, and optionally an ECU identifier, to a frame.

The default implementation `FrameMapDefault` represents the same behaviour as the DLT-Viewer.

The `FrameMapSimple` ignores the application and context identifier, which can be useful for checking that a FIBEX file has a unique message identifier for all files.

The `FrameMapEcu` and `FrameMapEcuSimple` extend the default and simple to optionally provide an ECU identifier, so multiple FIBEXes can be loaded for a single log file containing multiple ECUs.

Issue: DOTNET-737