DLT: Use NonVerboseByteDecoder as a Fallback
If there is an error in the length after decoding non-verbose, use the byte decoder as a fallback (as if no FIBEX file was provided).

Fix the INonVerboseDltDecoder so that it doesn't log, but sets the IDltLineBuilder with the error, so the main decoder logs. This allows the position of the error to be properly logged.

Issue: DOTNET-741