DltDump: Enable PCAPs to decode non-verbose with a FIBEX file
The `DltPcapTraceDecoder` now creates a factory, which returns a `IPcapTraceDecoder : ITraceDecoder<DltLineBase>`. This new interface exposes a way to set the time stamp. This replaces the mechanism where we passed around the `IOutputStream`. Now the factory has this as an internal detail along with the `IFrameMap` which is what we wanted to add. Arguably, passing the factory is better, as what we're doing is trying to create the trace reader on demand anyway. If we have other parameters or configuration for the decoder, we can continue to use the mechanism presented in this patch set.

Eventually, code gets to the `Connection` class which uses this factory to create the decoder for the payload portion of our PCAP packets.

We no longer distinguish between `IOutputStream` being `null` or not, and treat it just as an optional parameter. This simplifies the code slightly.

The setting of the time stamp is now optimised to set the time stamp on the `IDltLineBuilder`, where the time stamp is used to create the line. This incidentally removes the need to have an override, as the `DltLineBuilder` does what we want anyway.

Issue: DOTNET-746