DLT: Decode properly DLT lines with Extended Headers, missing ECUID
Add test cases for the DltTraceDecoder for Non-Verbose messages. Test when the DLT line has an ECU ID or not, an Extended Header or not. If there is an Extended header, the NOAR and MSIN fields for DLTv1 should be ignored (as this information comes from the FIBEX file).

The DltTraceDecoderBase should only decode the message info in an extended header for Control and Verbose messages. Otherwise the data comes from the FIBEX file. Fixed the case when a DLT message had an unknown MSIN, the non-verbose line would not decode and result in an error. The test class `DltPacketWriter.DltPacketBuilder.NonVerboseExt` sets the MSIN to zero for all non-verbose messages with an extended header that exposed this bug.

When decoding a DLT file with a storage header, the ECU field should be taken from: (1) the DLT line if available; or (2) from the FIBEX file. Fix the case that the ECUID in the Storage Header overrides the ECU ID always. When the `DltLineBuilder` sets the ECUID from the storage header, the feature flag was not set. So the `NonVerboseDltDecoder` should check the feature flags, instead of `null` for proper operation.

The rest are clean up in the test cases to allow better testing of Non-Verbose messages.

Issue: DOTNET-751