DLT: Provide correct arguments in case of Fallback
In case of decoding an error, or the wrong length when decoding non-verbose using a FIBEX file, the fallback was decoded, but the argument was appended, rather than it being reset and being added.

This resulted in lines having incorrect arguments, followed by the non-verbose argument at the end, which should have been the only argument.

Additionally, in case of errors when decoding, we would not fallback properly. Some cases there would be an error decoding that didn't set an argument error in the main decoder, that would result in spurious errors of wrong lengths, when it should have been logged as a decoder error.

Issue: DOTNET-751