DLT: Rework FrameMap for multiple ECU support
When an ECU ID is in the log file, but no FIBEX is found with this ECU-ID, previous behaviour defaulted to the first loaded. This is confusing when loading. Better is to indicate that such a message doesn't exist, similar with the same behaviour as the dlt-viewer behaviour when an APP/CTX identifier is not found.

Issue: DOTNET-752