DltDump: Fix time stamps when decoding PCAPs
Introduced with f2f749a ("DltDump: Remove unnecessary line override for the packet time stamp"), the first line in a UDP packet has the time stamp, subsequent lines do not, until there is a new UDP packet.

It occurs, because on every new line `DltLineBuilder.Reset()` is called, which clears the flag that there is a time stamp.

Now, we don't reset the time stamp, so if it is set once, it must be manually cleared for it to have a different time stamp.

Issue: DOTNET-756