DLT: NonVerbose treats messages with no APPID/CTXID/ECUID properly
With the introduction of 2ed6fbd ("DLT: Improve logging of DLT NonVerbose"), the default for the application, context and ECU id is an empty string, no longer `null`. This meant when getting a frame with no app/ctx/ecu identifier, it was considered the same as it being present, but empty, thus resulting in no match.

Make an explicit check that the fields are present before getting the map.

Issue: DOTNET-757