DLT Decoder: Fix GetLogInfoResponse Overflow to ignore COM if it isn't present
When reviewing `AUTOSAR_SWS_DiagnosticLogAndTrace_4.2.2.pdf`, it says:

> 9 – RESPONSE DATA OVERFLOW – If the generated response is to large for transmitting a single Dlt message (PDU is 65536 Byte) only Status type RESPONSE DATA OVERFLOW shall be send (No Application_IDs and com_interface entries are transmitted). Than the user could than decide to formulate a less complex request.

Newer specifications omit this, e.g. `AUTOSAR_PRS_LogAndTraceProtocol_release_1.4.0.pdf` says:

> 9 – RESPONSE DATA OVERFLOW – If the generated response is too large

The reduction in text is likely because the COM field is made "reserved" in later specifications, but it shouldn't affect the payload.

The best is to support both types. If there are some real world packets, then we can investigate further.

Issue: DOTNET-777